Provide the menu as a module (#1)
* Provide the menu as a module.

Split the menu functionality of main.py out into functions that can be imported by applications.
When run as main.py from the firmware, it runs the game selection menu as before. When imported, it exports these functions:
- menu: Synchronously runs the menu.
- menugen: Generator that allows background processing while the menu is running, in particular the list of menu entries may be modified live.

* Improved support for live changes to menu entries.

When the list of entries changes between iterations of the menu iterator:
- When the selected item moves: keep it selected at its new index, with uninterrupted animation.
- When the selected item disappears: select and display the item that is now at the same (clamped) index.

* Allow the initial selection to be specified as an optional argument.

Useful in a hierarchy of menus to return to a higher level at the place it was left.